front/requirements.txt
streamlit
pydantic
requests
pillow

python-dotenv

python-multipart